added second pass for scheme selection
