+++ lib/Analysis/SchemeSelectionAnalysis/BUILD
@@ -9,15 +9,16 @@ cc_library(
     hdrs = ["SchemeSelectionAnalysis.h"],
     deps = [
         "@heir//lib/Analysis:Utils",
+        "@heir//lib/Analysis/SchemeInfoAnalysis",
         "@heir//lib/Dialect/Mgmt/IR:Dialect",
         "@heir//lib/Dialect/Secret/IR:Dialect",
         "@llvm-project//llvm:Support",
         "@llvm-project//mlir:Analysis",
         "@llvm-project//mlir:ArithDialect",
-        "@llvm-project//mlir:MathDialect",
         "@llvm-project//mlir:CallOpInterfaces",
         "@llvm-project//mlir:FuncDialect",
         "@llvm-project//mlir:IR",
+        "@llvm-project//mlir:MathDialect",
         "@llvm-project//mlir:Support",
     ],
 )
