+++ lib/Transforms/AnnotateSchemeInfo/AnnotateSchemeInfo.cpp
@@ -1,5 +1,6 @@
 #include "lib/Transforms/AnnotateSchemeInfo/AnnotateSchemeInfo.h"
 
+#include "lib/Analysis/SchemeInfoAnalysis/SchemeInfoAnalysis.h"
 #include "lib/Analysis/SchemeSelectionAnalysis/SchemeSelectionAnalysis.h"
 #include "mlir/include/mlir/Analysis/DataFlow/ConstantPropagationAnalysis.h"  // from @llvm-project
 #include "mlir/include/mlir/Analysis/DataFlow/DeadCodeAnalysis.h"  // from @llvm-project
@@ -20,6 +21,7 @@ struct AnnotateSchemeInfo : impl::AnnotateSchemeInfoBase<AnnotateSchemeInfo> {
     DataFlowSolver solver;
     solver.load<dataflow::DeadCodeAnalysis>();
     solver.load<dataflow::SparseConstantPropagation>();
+    solver.load<SchemeInfoAnalysis>();
     solver.load<SchemeSelectionAnalysis>();
 
     auto result = solver.initializeAndRun(getOperation());
@@ -30,7 +32,8 @@ struct AnnotateSchemeInfo : impl::AnnotateSchemeInfoBase<AnnotateSchemeInfo> {
       return;
     }
 
-    annotateNatureOfComputation(getOperation(), &solver, verbose);
+    annotateNatureOfComputation(getOperation(), &solver);
+    annotateModuleWithScheme(getOperation(), &solver);
   }
 };
